fix bug bound turn, add grid plot
