+++ Project.toml
@@ -5,9 +5,13 @@ version = "0.1.0"
 
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
+Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
 GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
+GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
 InfrastructureModels = "2030c09a-7f63-5d83-885d-db604e0e9cc0"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
@@ -33,3 +37,7 @@ Clp = "e2554f3b-3117-50c0-817c-e040a3ddf72d"
 Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+GraphPlot = "a2cc645c-3eea-5389-862e-a155d0052231"
+Fontconfig = "186bb1d3-e1f7-5a2c-a377-96d770f13627"
+Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
+Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
